+++ Source/Common/JUCEFileIO.h
@@ -5,9 +5,9 @@
 #include <assert.h>
 
 // JUCE FILE IO
-extern void getWavFileInfo(String path, unsigned* numFrames, double* sampleRate);
+extern void getWavFileInfo(String path, unsigned* numFrames, float* sampleRate);
 extern void loadWavFile(String path, unsigned chunkSize, unsigned numFrames, float* buffer);
-extern void saveWavFile(String path, String desc, String origin, double sampleRate, unsigned chunkSize, unsigned numFrames, float* data);
+extern void saveWavFile(String path, String desc, String origin, float sampleRate, unsigned chunkSize, unsigned numFrames, float* data);
 extern void saveTextFile(String path, String text);
 
 #endif
